Save DM Bug

+++ src/routes/dms/[dmId]/+page.server.ts
@@ -22,19 +22,22 @@ export const actions = {
 	saveDM: async (event) => {
 		const session = await event.locals.getSession();
 		if (!session?.user) throw redirect(301, "/");
+		if (!event.params.dmId) throw redirect(301, "/dms");
 
-		const data = await event.request.formData();
-		const dmId = (data.get("dmId") || "") as string;
-
-		const dm = await getUserDMWithLogs(session.user.id, dmId);
+		const dm = await getUserDMWithLogs(session.user.id, event.params.dmId);
 		if (!dm) throw redirect(301, "/dms");
 
-		if (dm.logs.length) throw redirect(301, `/dms/${dmId}`);
+		const data = await event.request.formData();
+		const result = await saveDM(event.params.dmId, session.user.id, {
+			id: dm.id,
+			name: data.get("name") as string,
+			DCI: (data.get("DCI") as string) || null,
+			uid: dm.uid || ""
+		});
 
-		dm.name = data.get("name") as string;
-		dm.DCI = data.get("DCI") as string;
+		if (result && result.id) throw redirect(301, `/dms`);
 
-		return await saveDM(dmId, session.user.id, dm);
+		return result;
 	},
 	deleteDM: async (event) => {
 		const session = await event.locals.getSession();
@@ -45,6 +48,8 @@ export const actions = {
 
 		const dm = await getUserDMWithLogs(session.user.id, dmId);
 		if (!dm) throw redirect(301, "/dms");
+
+		if (dm.logs.length) return { id: null, error: "You cannot delete a DM that has logs" };
 
 		return await deleteDM(dmId, session.user.id);
 	}
